Added a timer for the bossRoom
Made the player unable to leave the boss room until they start the timer and the timer is done counting down.
added a Key class as button and a new sprite for the door to show it is locked.
Added a visible timer to show how much time is left.